requirements.txt
streamlit==1.31.1
openai==1.11.1
PyPDF2==3.0.1
python-dotenv==1.0.1
typing-extensions==4.9.0
protobuf==3.20.3
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz